scripts/requirements.txt
boto3>=1.26.0
python-dotenv>=0.21.0
AWSIoTPythonSDK>=1.4.9
requests>=2.28.0